src/test/java/com/artipie/http/headers/AuthorizationTest.java
/*
 * MIT License
 *
 * Copyright (c) 2020 Artipie
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
package com.artipie.http.headers;

import com.artipie.http.Headers;
import org.hamcrest.MatcherAssert;
import org.hamcrest.core.IsEqual;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

/**
 * Test case for {@link Authorization}.
 *
 * @since 0.12
 */
public final class AuthorizationTest {

    @Test
    void shouldHaveExpectedName() {
        MatcherAssert.assertThat(
            new Authorization("Basic abc").getKey(),
            new IsEqual<>("Authorization")
        );
    }

    @Test
    void shouldHaveExpectedValue() {
        final String value = "Basic 123";
        MatcherAssert.assertThat(
            new Authorization(value).getValue(),
            new IsEqual<>(value)
        );
    }

    @Test
    void shouldExtractValueFromHeaders() {
        final String value = "Bearer abc";
        final Authorization header = new Authorization(
            new Headers.From(
                new Header("Content-Length", "11"),
                new Header("authorization", value),
                new Header("X-Something", "Some Value")
            )
        );
        MatcherAssert.assertThat(header.getValue(), new IsEqual<>(value));
    }

    @Test
    void shouldHaveExpectedScheme() {
        MatcherAssert.assertThat(
            new Authorization("Digest abc===").scheme(),
            new IsEqual<>("Digest")
        );
    }

    @Test
    void shouldHaveExpectedCredentials() {
        MatcherAssert.assertThat(
            new Authorization("Bearer 123.abc").credentials(),
            new IsEqual<>("123.abc")
        );
    }

    @Test
    void shouldFailToExtractValueFromEmptyHeaders() {
        Assertions.assertThrows(
            IllegalStateException.class,
            () -> new Authorization(Headers.EMPTY).getValue()
        );
    }

    @Test
    void shouldFailToExtractValueWhenNoAuthorizationHeaders() {
        Assertions.assertThrows(
            IllegalStateException.class,
            () -> new Authorization(
                new Headers.From("Content-Type", "text/plain")
            ).getValue()
        );
    }

    @Test
    void shouldFailToExtractValueFromMultipleHeaders() {
        Assertions.assertThrows(
            IllegalStateException.class,
            () -> new Authorization(
                new Headers.From(
                    new Authorization("Bearer one"),
                    new Authorization("Bearer two")
                )
            ).getValue()
        );
    }
}
